Tweak docs, split credentials into two parts
These updates are inspired by some sessions where I paired with devs
while they added gha-scala-library-release-workflow to an existing
repo, working from the existing docs.

When adding the workflow to an existing repo, it's mostly the case
that existing release credentials are being reused - instructions on
how to create _new_ release credentials are unnecessary and distracting.
It makes sense to split the docs for that off into a separate file.