Explicitly switch to Sonatype token authentication
As of January 2024, Sonatype is actively discouraging the legacy
username & password method of authentication, recommending token
authentication instead:

* https://central.sonatype.org/news/20240109_issues_sonatype_org_deprecation/#support-requests
* https://central.sonatype.org/publish/generate-token/

In this new scheme, the token is still split into a username/password
format, and both are randomised strings, making the username portion
a meaningful secret (ie one that can be revoked) and so worthy of being
treated as a secret. Consequently, in this change username is now a
GitHub workflow 'secret' parameter, rather than a simple input.