Disable default behaviour of base64 wrapping at 76 columns
While attempting to create a preview release of a branch, Fred encountered the rather cryptic error:

```
Error: Invalid format 'bC1tZXJzaW9uLjIwMjQtMDMtMDVUMTMyNC5mYzMyOWVmNy1TTkFQU0hPVCIK
```

With some additional logging, we can see that the GITHUB_OUTPUT file contains the malformed output:

```
version_file_post_release_content=VGhpc0J1aWxkIC8gdmVyc2lvbiA6PSAiMjMuMC4xLVBSRVZJRVcucmVjaXBlLXNjaGVtYS1tb2Rl
bC1tZXJzaW9uLjIwMjQtMDMtMDVUMTMyNC5mYzMyOWVmNy1TTkFQU0hPVCIK
```

This is because `base64` on Ubuntu defaults to line wrapping at 76 characters. This behaviour is different from `base64` on MacOS, which does not wrap.

It seemed strange that we did not see this error on other branches, but this can be explained by the knowledge that base64 encoding uses an equals sign to indicate padding if the input string length is not a multiple of 3 bytes.

So for a branch name that is not a multiple of 3 bytes, the GITHUB_OUTPUT file would contain something like:
```
version_file_post_release_content=VGhpc0J1aWxkIC8gdmVyc2lvbiA6PSAiMjMuMC4wLVBSRVZJRVcubGlzdC1pdGVtLXR5cGUuMjAy
NC0wMy0wNVQxMjUwLjk3MzY2NmVlIgo=
```

The second line in this excerpt is nonsense, and the `version_file_post_release_content` variable will be truncated. But both lines are parseable, so the release completes successfully.

The bug can be fixed by adding `-w0` to the `base64` command. This tells `base64` not to line wrap.