Fix link to create GitHub App: use `pull_requests` with underscore
This fixes a small documentation mistake I made in https://github.com/guardian/gha-scala-library-release-workflow/pull/26,
where usage of a Github App was introduced. I supplied a link in the docs to help create a new GitHub App, this link was
supposed to pre-populate the form with the 3 necessary permissions (Contents: Read and write, Pull Requests: Read and write,
and Metadata: Read-only). However, checking it now, it only provides 2 of 3, and the 'Pull Requests' permission wasn't
set.

It turned out that I should have been using `pull_requests=write` in the url, rather than `pull-requests=write` - this
isn't documented, so far as I can see:

https://docs.github.com/en/apps/sharing-github-apps/registering-a-github-app-using-url-parameters

...but in the end I managed to guess it!